push simple model

+++ mmdet/models/losses/.ipynb_checkpoints/utils-checkpoint.py
@@ -37,7 +37,7 @@ def weight_reduce_loss(loss, weight=None, reduction='mean', avg_factor=None):
     """
     # if weight is specified, apply element-wise weight
     if weight is not None:
-        loss = loss * weight
+        loss = loss * tf.cast(weight,loss.dtype)
 
     # if avg_factor is not specified, just reduce the loss
     if avg_factor is None:
